chore: refactor layout to keep features and configurations in a per-project directory
