+++ e2e/configurations/metricbeat/mysql.yml
