+++ e2e/configurations/metricbeat/redis.yml
