+++ e2e/configurations/metricbeat/vsphere.yml
