+++ e2e/elasticsearch.go
@@ -1,4 +1,4 @@
-package metricbeat
+package e2e
 
 import (
 	"bytes"
