+++ e2e/features/metricbeat/redis.feature
