+++ e2e/test_utils.go
@@ -1,4 +1,4 @@
-package metricbeat
+package e2e
 
 import (
 	"io"
